+++ std.go
@@ -89,6 +89,10 @@ go.build = function build() {
   return go.apply(null, ['build'].concat(Array.prototype.slice.call(arguments)));
 };
 
+go.fix = function fix() {
+  return go.apply(null, ['fix'].concat(Array.prototype.slice.call(arguments)));
+};
+
 go.generate = function generate() {
   return go.apply(null, ['generate'].concat(Array.prototype.slice.call(arguments)));
 };