+++ CHANGES.rst
@@ -9,6 +9,7 @@ Version 0.3
   * Fix ``dep.prune`` function.
   * Add ``go.build`` function.
   * Add ``go.fix`` function.
+  * Add ``go.fmt`` function.
 
 
 Version 0.2